Added base class for OptimizationProblem
